Lint: Java 5 migration

+++ app/src/main/java/com/csvreader/CsvReader.java
@@ -1405,7 +1405,7 @@ public class CsvReader {
         Object indexValue = headersHolder.IndexByName.get(headerName);
 
         if (indexValue != null) {
-            return ((Integer) indexValue).intValue();
+            return (Integer) indexValue;
         } else {
             return -1;
         }
